Revert "Read SourceLink info and call service to retrieve source from there (#57978)"
This reverts commit f17419c5f71c744d07e9d15b9312cc7042983f9b.
