+++ src/Features/Core/Portable/PdbSourceDocument/DocumentDebugInfoReader.cs
@@ -8,7 +8,6 @@ using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
 using Microsoft.CodeAnalysis.Debugging;
 using Microsoft.CodeAnalysis.PooledObjects;
-using Microsoft.SourceLink.Tools;
 
 namespace Microsoft.CodeAnalysis.PdbSourceDocument
 {
@@ -48,42 +47,11 @@ namespace Microsoft.CodeAnalysis.PdbSourceDocument
                 var checksum = _pdbReader.GetBlobContent(document.Hash);
 
                 var embeddedTextBytes = TryGetEmbeddedTextBytes(handle);
-                var sourceLinkUrl = TryGetSourceLinkUrl(handle);
 
-                sourceDocuments.Add(new SourceDocument(filePath, hashAlgorithm, checksum, embeddedTextBytes, sourceLinkUrl));
+                sourceDocuments.Add(new SourceDocument(filePath, hashAlgorithm, checksum, embeddedTextBytes));
             }
 
             return sourceDocuments.ToImmutable();
-        }
-
-        private string? TryGetSourceLinkUrl(DocumentHandle handle)
-        {
-            var document = _pdbReader.GetDocument(handle);
-            if (document.Name.IsNil)
-                return null;
-
-            var documentName = _pdbReader.GetString(document.Name);
-            if (documentName is null)
-                return null;
-
-            foreach (var cdiHandle in _pdbReader.GetCustomDebugInformation(EntityHandle.ModuleDefinition))
-            {
-                var cdi = _pdbReader.GetCustomDebugInformation(cdiHandle);
-                if (_pdbReader.GetGuid(cdi.Kind) == PortableCustomDebugInfoKinds.SourceLink && !cdi.Value.IsNil)
-                {
-                    var blobReader = _pdbReader.GetBlobReader(cdi.Value);
-                    var sourceLinkJson = blobReader.ReadUTF8(blobReader.Length);
-
-                    var map = SourceLinkMap.Parse(sourceLinkJson);
-
-                    if (map.TryGetUri(documentName, out var uri))
-                    {
-                        return uri;
-                    }
-                }
-            }
-
-            return null;
         }
 
         private byte[]? TryGetEmbeddedTextBytes(DocumentHandle handle)
