+++ src/Features/Core/Portable/PdbSourceDocument/PdbFileLocatorService.cs
@@ -4,9 +4,7 @@
 
 using System;
 using System.Composition;
-using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,29 +17,23 @@ namespace Microsoft.CodeAnalysis.PdbSourceDocument
     [Export(typeof(IPdbFileLocatorService)), Shared]
     internal sealed class PdbFileLocatorService : IPdbFileLocatorService
     {
-        private const int SymbolLocatorTimeout = 2000;
-
-        private readonly ISourceLinkService? _sourceLinkService;
-
         [ImportingConstructor]
-        [SuppressMessage("RoslynDiagnosticsReliability", "RS0033:Importing constructor should be [Obsolete]", Justification = "Used in test code")]
-        public PdbFileLocatorService([Import(AllowDefault = true)] ISourceLinkService? sourceLinkService)
+        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+        public PdbFileLocatorService()
         {
-            _sourceLinkService = sourceLinkService;
         }
 
-        public async Task<DocumentDebugInfoReader?> GetDocumentDebugInfoReaderAsync(string dllPath, IPdbSourceDocumentLogger? logger, CancellationToken cancellationToken)
+        public Task<DocumentDebugInfoReader?> GetDocumentDebugInfoReaderAsync(string dllPath, CancellationToken cancellationToken)
         {
             var dllStream = IOUtilities.PerformIO(() => File.OpenRead(dllPath));
             if (dllStream is null)
-                return null;
+                return Task.FromResult<DocumentDebugInfoReader?>(null);
 
             Stream? pdbStream = null;
             DocumentDebugInfoReader? result = null;
             var peReader = new PEReader(dllStream);
             try
             {
-                // Try to load the pdb file from disk, or embedded
                 if (peReader.TryOpenAssociatedPortablePdb(dllPath, pdbPath => File.OpenRead(pdbPath), out var pdbReaderProvider, out _))
                 {
                     Contract.ThrowIfNull(pdbReaderProvider);
@@ -49,34 +41,17 @@ namespace Microsoft.CodeAnalysis.PdbSourceDocument
                     result = new DocumentDebugInfoReader(peReader, pdbReaderProvider);
                 }
 
-                // Otherwise call the debugger to find the PDB from a symbol server etc.
-                if (result is null && _sourceLinkService is not null)
+                // TODO: Otherwise call the debugger to find the PDB from a symbol server etc.
+                if (result is null)
                 {
-                    var delay = Task.Delay(SymbolLocatorTimeout, cancellationToken);
-                    var pdbResultTask = _sourceLinkService.GetPdbFilePathAsync(dllPath, peReader, logger, cancellationToken);
-
-                    var winner = await Task.WhenAny(pdbResultTask, delay).ConfigureAwait(false);
-
-                    if (winner == pdbResultTask)
-                    {
-                        var pdbResult = await pdbResultTask.ConfigureAwait(false);
-
-                        // TODO: Support windows PDBs: https://github.com/dotnet/roslyn/issues/55834
-                        // TODO: Log results from pdbResult.Log: https://github.com/dotnet/roslyn/issues/57352
-                        if (pdbResult is not null && pdbResult.IsPortablePdb)
-                        {
-                            pdbStream = IOUtilities.PerformIO(() => File.OpenRead(pdbResult.PdbFilePath));
-                            if (pdbStream is not null)
-                            {
-                                var readerProvider = MetadataReaderProvider.FromPortablePdbStream(pdbStream);
-                                result = new DocumentDebugInfoReader(peReader, readerProvider);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        // TODO: Log the timeout: https://github.com/dotnet/roslyn/issues/57352
-                    }
+                    // Debugger needs:
+                    // - PDB MVID
+                    // - PDB Age
+                    // - PDB TimeStamp
+                    // - PDB Path
+                    // - DLL Path
+                    // 
+                    // Most of this info comes from the CodeView Debug Directory from the dll
                 }
             }
             catch (BadImageFormatException)
@@ -96,7 +71,7 @@ namespace Microsoft.CodeAnalysis.PdbSourceDocument
                 }
             }
 
-            return result;
+            return Task.FromResult<DocumentDebugInfoReader?>(result);
         }
     }
 }
